:memo: refactor: elements and styles
