--- README.md
@@ -1,49 +1,0 @@
-<h1 align="center">
-  📖 Dt-Money
-</h1>
-
-<p align="center">
-  <img alt="GitHub top language" src="https://img.shields.io/github/languages/top/manoguii/dt-money?color=blue">
-  <img alt="Repository size" src="https://img.shields.io/github/repo-size/manoguii/dt-money?color=blue">
-  <img alt="GitHub language count" src="https://img.shields.io/github/languages/count/manoguii/dt-money?color=blue">
-</p>
-
-<div align="left">
-  <img src="assets/dt-money.png" alt="faladev" >
-</div>
-
-## Instalação
-
-- Clone o repositório
-```bash
-$ git clone https://github.com/manoguii/dt-money.git
-```
-
-- Instale as dependências
-```bash
-$ npm install
-```
-
-- Execute o banco de dados
-```bash
-$ npm run dev:server
-```
-
-- Execute a aplicação
-```bash
-$ npm run dev
-```
-
-## Sobre
-
-O dt-money é um projeto para gerenciar despesas e receitas. 🚀
-
-
-## Tecnologias utilizadas 👩🏻‍💻
-
-- [ReactJS](https://reactjs.org/)
-- [TypeScript](https://www.typescriptlang.org/)
-
----
-
-<p align="center">Made with 💙 by Guilherme David</p>
