+++ package.json
@@ -6,30 +6,27 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "preview": "vite preview",
-    "dev:server": "json-server server.json -w -d 500"
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@hookform/resolvers": "^2.9.8",
-    "@radix-ui/react-dialog": "^1.0.0",
-    "@radix-ui/react-radio-group": "^1.0.0",
-    "axios": "^0.27.2",
-    "phosphor-react": "^1.4.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.36.1",
-    "styled-components": "^5.3.5",
-    "zod": "^3.19.1"
+    "@hookform/resolvers": "3.0.1",
+    "@radix-ui/react-dialog": "1.0.3",
+    "@radix-ui/react-radio-group": "1.1.2",
+    "axios": "1.3.5",
+    "phosphor-react": "1.4.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-hook-form": "7.43.9",
+    "styled-components": "5.3.9",
+    "zod": "3.21.4"
   },
   "devDependencies": {
-    "@rocketseat/eslint-config": "^1.1.3",
-    "@types/react": "^18.0.17",
-    "@types/react-dom": "^18.0.6",
-    "@types/styled-components": "^5.1.26",
-    "@vitejs/plugin-react": "^2.1.0",
-    "eslint": "^8.24.0",
-    "json-server": "^0.17.0",
-    "typescript": "^4.6.4",
-    "vite": "^3.1.0"
+    "@rocketseat/eslint-config": "^1.2.0",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@types/styled-components": "5.1.26",
+    "@vitejs/plugin-react": "^3.1.0",
+    "typescript": "^4.9.3",
+    "vite": "^4.2.0"
   }
 }
