--- server.json
@@ -1,76 +1,0 @@
-{
-  "transactions": [
-    {
-      "id": 1,
-      "description": "Curso Ignite",
-      "type": "outcome",
-      "category": "Educação",
-      "price": 1299,
-      "created_at": "2015-02-14T19:23:21Z"
-    },
-    {
-      "id": 2,
-      "description": "Pizza",
-      "type": "outcome",
-      "category": "Alimentação",
-      "price": 39,
-      "created_at": "2015-02-09T19:23:21Z"
-    },
-    {
-      "id": 3,
-      "description": "Pagamento",
-      "type": "income",
-      "category": "Pagamento",
-      "price": 2400,
-      "created_at": "2015-02-01T19:23:21Z"
-    },
-    {
-      "description": "Desenvolvimento de app",
-      "price": 1522,
-      "category": "Educação",
-      "type": "outcome",
-      "created_at": "2022-09-27T23:46:43.329Z",
-      "id": 4
-    },
-    {
-      "description": "Pastel",
-      "price": 5,
-      "category": "Alimentação",
-      "type": "outcome",
-      "created_at": "2022-09-28T00:04:48.159Z",
-      "id": 5
-    },
-    {
-      "description": "Salario",
-      "price": 880,
-      "category": "Adiantamento",
-      "type": "income",
-      "created_at": "2022-09-28T00:05:29.871Z",
-      "id": 6
-    },
-    {
-      "description": "Curso",
-      "price": 230,
-      "category": "Estudos",
-      "type": "income",
-      "created_at": "2022-10-23T22:34:08.089Z",
-      "id": 7
-    },
-    {
-      "description": "Patel",
-      "price": 12,
-      "category": "Alimentação",
-      "type": "outcome",
-      "created_at": "2022-10-28T23:08:52.644Z",
-      "id": 8
-    },
-    {
-      "description": "PG",
-      "price": 1200,
-      "category": "Salario",
-      "type": "income",
-      "created_at": "2022-10-28T23:09:32.411Z",
-      "id": 9
-    }
-  ]
-}
