+++ src/@types/styles.d.ts
@@ -1,5 +1,5 @@
 import 'styled-components'
-import { defaultTheme } from '../styles/themes/default'
+import { defaultTheme } from '../styles/theme/default'
 
 type ThemeType = typeof defaultTheme
