+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { ThemeProvider } from 'styled-components'
-import { TransactionsProvider } from './contexts/TransactionsContexts'
-import { Transactions } from './pages/Transactions'
 import { GlobalStyle } from './styles/global'
-import { defaultTheme } from './styles/themes/default'
+import { defaultTheme } from './styles/theme/default'
+import { Transactions } from './pages/Transactions'
+import { TransactionsProvider } from './contexts/TransactionsContexts'
 
 export function App() {
   return (
