+++ src/components/Summary/index.tsx
@@ -1,6 +1,6 @@
 import { SummaryCard, SummaryContainer } from './styles'
 import { ArrowCircleDown, ArrowCircleUp, CurrencyDollar } from 'phosphor-react'
-import { priceFormated } from '../../utils/Formatter'
+import { priceFormatted } from '../../utils/Formatter'
 import { useSummary } from '../../hooks/useSummary'
 
 export function Summary() {
@@ -14,23 +14,25 @@ export function Summary() {
           <ArrowCircleUp size={32} color="#00b37e" />
         </header>
 
-        <strong>{priceFormated.format(summary.income)}</strong>
+        <strong>{priceFormatted.format(summary.income)}</strong>
       </SummaryCard>
+
       <SummaryCard>
         <header>
           <span>Despesas</span>
           <ArrowCircleDown size={32} color="#f75a68" />
         </header>
 
-        <strong>{priceFormated.format(summary.outcome)}</strong>
+        <strong>{priceFormatted.format(summary.outcome)}</strong>
       </SummaryCard>
+
       <SummaryCard variant="green">
         <header>
           <span>Total</span>
           <CurrencyDollar size={32} color="#fff" />
         </header>
 
-        <strong>{priceFormated.format(summary.total)}</strong>
+        <strong>{priceFormatted.format(summary.total)}</strong>
       </SummaryCard>
     </SummaryContainer>
   )
