+++ src/components/Summary/styles.ts
@@ -19,9 +19,8 @@ interface SummaryCardProps {
 
 export const SummaryCard = styled.div<SummaryCardProps>`
   background: ${(props) => props.theme['gray-600']};
-  border-radius: 25px;
+  border-radius: 6px;
   padding: 1.5rem;
-  align-self: center;
 
   header {
     display: flex;
