+++ src/pages/Transactions/index.tsx
@@ -2,7 +2,7 @@ import { useContext } from 'react'
 import { Header } from '../../components/Header'
 import { Summary } from '../../components/Summary'
 import { TransactionContext } from '../../contexts/TransactionsContexts'
-import { dateFormated, priceFormated } from '../../utils/Formatter'
+import { dateFormatted, priceFormatted } from '../../utils/Formatter'
 import { SearchForm } from './components/SearchForm'
 import {
   PriceHighlight,
@@ -12,6 +12,7 @@ import {
 
 export function Transactions() {
   const { transactions } = useContext(TransactionContext)
+
   return (
     <div>
       <Header />
@@ -29,12 +30,12 @@ export function Transactions() {
                   <td>
                     <PriceHighlight variant={transaction.type}>
                       {transaction.type === 'outcome' && '- '}
-                      {priceFormated.format(transaction.price)}
+                      {priceFormatted.format(transaction.price)}
                     </PriceHighlight>
                   </td>
                   <td>{transaction.category}</td>
                   <td>
-                    {dateFormated.format(new Date(transaction.created_at))}
+                    {dateFormatted.format(new Date(transaction.created_at))}
                   </td>
                 </tr>
               )
