+++ src/styles/global.ts
@@ -17,8 +17,8 @@ export const GlobalStyle = createGlobalStyle`
     color: ${(props) => props.theme['gray-100']};
     -webkit-font-smoothing: antialiased;
   }
-
-  body, input-security, textarea, button {
+  
+  body, input, textarea, button {
     font: 400 1rem Roboto, sans-serif;
   }
 `
