+++ src/utils/Formatter.tsx
@@ -1,5 +1,6 @@
-export const dateFormated = new Intl.DateTimeFormat('pt-BR')
-export const priceFormated = new Intl.NumberFormat('pt-BR', {
+export const dateFormatted = new Intl.DateTimeFormat('pt-BR')
+
+export const priceFormatted = new Intl.NumberFormat('pt-BR', {
   style: 'currency',
   currency: 'BRL',
 })
